feat(build): support Linux aarch64 builds
The architecture-specific build logic keyed on PAIMON_CPU_FLAG and
PAIMON_ARMV8_ARCH, neither of which was ever defined, so aarch64 got no
-march= at all. TargetArchitecture.cmake now resolves the target
architecture once, and PAIMON_AARCH64_MARCH selects the -march= value
(default armv8-a; an empty value passes no -march flag). Fixing the
target-dependent behavior this exposed changes one result on x86-64 as
well: out of range float to integer conversion now follows Java.

- Remove dead SIMD build logic: the SSE4.2 CRC32C kernel (Castagnoli,
  never selected, not the persisted zlib CRC-32, so no on-disk value
  changes), the ppc AltiVec probe, and the unread PAIMON_HAVE_NEON /
  ARMV8_CRC / ARMV8_CRYPTO definitions. crc32c_test.cpp now pins the
  checksum to zlib.crc32 reference values.
- Name the package after the host platform, e.g.
  paimon-cpp-linux-aarch64.tar.gz; add --platform and --print-name to
  build_and_package.sh. Lumina is prebuilt for linux-x86_64 only and
  now fails configuration elsewhere with an explicit error.
- Run the same CI matrix on x86_64 and aarch64 (clang-tidy excepted),
  plus a script-tests job for the new CMake-module, packaging-argument
  and asan_symbolize tests; all jobs are required checks.
- Fix char-signedness defects the unsigned-char AArch64 Linux ABI
  exposed: BinaryString UTF-8 leading-byte classification (NumChars,
  Substring and IndexOf counted bytes, not characters), tolower and
  toupper on bytes at or above 0x80, TINYINT min/max aggregation
  (min(-20, 10) returned 10) and VariantValueToString (-20 printed as
  236).
- Replace the undefined float/double to integer static_casts, in the
  cast executor and in the Arrow kernel it called, with
  JavaFloatingToIntegerCast: NaN to 0, out of range saturating at the
  int32/int64 bounds, then narrowing to width. Both the literal and
  the array path use it so stats and data stay consistent.
- Link compiler-rt's builtins into the Clang UBSan build: the 128-bit
  multiplication overflow check calls __muloti4, which libgcc does not
  provide and Clang does not inline on aarch64, so the sanitizer build
  did not link there.
- Pin asan_symbolize.py stdin/stdout to UTF-8 with surrogateescape so
  a byte that is not valid UTF-8 no longer truncates the test log.
- docs: a supported platform matrix, rules on char signedness and on
  float to integer conversion, and Java-consistent results in the type
  change support matrix.

+++ src/paimon/common/utils/binary_row_partition_computer_test.cpp
@@ -77,7 +77,9 @@ TEST(BinaryRowPartitionComputerTest, TestToAndFromBinaryRow) {
         ASSERT_OK_AND_ASSIGN(BinaryRow row, computer->ToBinaryRow(partition_map));
         ASSERT_EQ(12, row.GetFieldCount());
         ASSERT_EQ(true, row.GetBoolean(0));
-        ASSERT_EQ(-20, row.GetByte(1));
+        // GetByte() returns a plain char, whose signedness the target ABI decides and
+        // -fsigned-char / -funsigned-char can flip; the AArch64 Linux ABI makes it unsigned.
+        ASSERT_EQ(-20, static_cast<int8_t>(row.GetByte(1)));
         ASSERT_EQ(10, row.GetByte(2));
         ASSERT_EQ(1556, row.GetShort(3));
         ASSERT_EQ(-2556, row.GetShort(4));
@@ -121,7 +123,7 @@ TEST(BinaryRowPartitionComputerTest, TestToAndFromBinaryRow) {
         ASSERT_OK_AND_ASSIGN(BinaryRow row, computer->ToBinaryRow(partition_map));
         ASSERT_EQ(12, row.GetFieldCount());
         ASSERT_EQ(true, row.GetBoolean(0));
-        ASSERT_EQ(-20, row.GetByte(1));
+        ASSERT_EQ(-20, static_cast<int8_t>(row.GetByte(1)));
         ASSERT_EQ(10, row.GetByte(2));
         ASSERT_EQ(1556, row.GetShort(3));
         ASSERT_EQ(-2556, row.GetShort(4));
@@ -165,7 +167,7 @@ TEST(BinaryRowPartitionComputerTest, TestToAndFromBinaryRow) {
         ASSERT_OK_AND_ASSIGN(BinaryRow row, computer->ToBinaryRow(partition_map));
         ASSERT_EQ(12, row.GetFieldCount());
         ASSERT_EQ(true, row.GetBoolean(0));
-        ASSERT_EQ(-20, row.GetByte(1));
+        ASSERT_EQ(-20, static_cast<int8_t>(row.GetByte(1)));
         ASSERT_EQ(10, row.GetByte(2));
         ASSERT_EQ(1556, row.GetShort(3));
         ASSERT_EQ(-2556, row.GetShort(4));
