+++ src/paimon/common/utils/data_converter_utils_test.cpp
@@ -20,6 +20,7 @@
 #include "paimon/common/utils/data_converter_utils.h"
 
 #include <cstddef>
+#include <cstdint>
 #include <memory>
 #include <vector>
 
@@ -71,7 +72,9 @@ TEST(DataConverterUtilsTest, TestDataToBinaryRowConverterWithLegacyPartitionName
     ASSERT_EQ(data.size(), row.GetFieldCount());
     ASSERT_EQ(true, row.GetBoolean(0));
     ASSERT_EQ(10, row.GetByte(1));
-    ASSERT_EQ(-20, row.GetByte(2));
+    // GetByte() returns a plain char, whose signedness the target ABI decides and
+    // -fsigned-char / -funsigned-char can flip; the AArch64 Linux ABI makes it unsigned.
+    ASSERT_EQ(-20, static_cast<int8_t>(row.GetByte(2)));
     ASSERT_EQ(1556, row.GetShort(3));
     ASSERT_EQ(-2556, row.GetShort(4));
     ASSERT_EQ(348489, row.GetInt(5));
@@ -125,7 +128,9 @@ TEST(DataConverterUtilsTest, TestDataToBinaryRowConverterWithNoLegacyPartitionNa
     ASSERT_EQ(data.size(), row.GetFieldCount());
     ASSERT_EQ(true, row.GetBoolean(0));
     ASSERT_EQ(10, row.GetByte(1));
-    ASSERT_EQ(-20, row.GetByte(2));
+    // GetByte() returns a plain char, whose signedness the target ABI decides and
+    // -fsigned-char / -funsigned-char can flip; the AArch64 Linux ABI makes it unsigned.
+    ASSERT_EQ(-20, static_cast<int8_t>(row.GetByte(2)));
     ASSERT_EQ(1556, row.GetShort(3));
     ASSERT_EQ(-2556, row.GetShort(4));
     ASSERT_EQ(348489, row.GetInt(5));
